+++ PSUtil/internal/scripts/preimport.ps1
@@ -1,4 +1,4 @@
-# Add all things you want to run before importing the main code
+﻿# Add all things you want to run before importing the main code
 foreach ($function in (Get-ChildItem (Join-PSFPath $script:ModuleRoot 'internal' 'configurations' 'validations')))
 {
 	. Import-ModuleFile -Path $function.FullName